refactor: add check productImage query

+++ src/main/java/com/trade_accounting/services/impl/ImageServiceImpl.java
@@ -47,15 +47,9 @@ public class ImageServiceImpl implements ImageService {
     @Override
     public void deleteById(Long id) {
         Optional<Image> optional = imageRepository.findById(id);
-        if (optional.isPresent()) {
-            try {
-                imageRepository.deleteById(id);
-                if(!imageRepository.existsById(id)) {
-                    Files.deleteIfExists(Paths.get(optional.get().getImageUrl()));
-                }
-            } catch (Exception ignored) {
-
-            }
+        if (optional.isPresent() && imageRepository.countProductImage(id) == 0) {
+            imageRepository.deleteById(id);
+            Files.deleteIfExists(Paths.get(optional.get().getImageUrl()));
         }
     }
